1a/requirements.txt
marker-pdf==1.8.2
torch>=2.0.0